src/requirements.txt
ffmpy
pydub
python-docx
functional